More styling fixes. Added question and answer notifications
