+++ src/Visualization/Components/Styles/HomePage.css
@@ -5,84 +5,6 @@
 	min-height: 80vh;
 }
 
-.home-page .home-page-showcase > label {
-	font-family: 'Lucida Console', 'Courier New', monospace;
-	color: black;
-	width: 70%;
-	text-transform: uppercase;
-	font-size: 25px;
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-start;
-	margin-top: 50px;
-	margin-left: 100px;
-}
-
-.home-page .home-page-showcase .product-showcase {
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-	margin-bottom: 20px;
-	margin-top: 10px;
-}
-
-.home-page .home-page-showcase .product-showcase-product {
-	position: relative;
-	display: flex;
-	flex-direction: row;
-	padding: 0.5%;
-	border: 2px solid #7dd359;
-	border-radius: 5px;
-	background-color: #c5c5c5;
-	width: 450px;
-	margin: 0 15px;
-}
-
-.home-page .home-page-showcase .product-showcase-info {
-	display: flex;
-	flex-direction: column;
-}
-
-.home-page .home-page-showcase .product-showcase-product:hover {
-	background-color: #b1b1b1;
-	cursor: pointer;
-}
-
-.home-page .home-page-showcase .product-showcase-product img {
-	height: 200px;
-	width: 200px;
-	margin-right: 10px;
-}
-
-.home-page .home-page-showcase .product-showcase-product label {
-	text-transform: uppercase;
-	width: 100%;
-	font-family: 'Merriweather Sans', sans-serif;
-	font-size: medium;
-}
-
-.home-page .home-page-showcase .product-showcase-product p {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-weight: lighter;
-}
-
-.home-page .home-page-showcase .product-showcase-product button {
-	position: absolute;
-	padding: 2%;
-	font-family: 'Merriweather Sans', sans-serif;
-	border-radius: 4px;
-	background-color: #7dd359;
-	color: black;
-	width: 100px;
-	right: 2%;
-	bottom: 2%;
-}
-
-.home-page .home-page-showcase .product-showcase-product button:hover {
-	background-color: rgb(35, 216, 133);
-	cursor: pointer;
-}
-
 .home-page .product-grid {
 	display: flex;
 	flex-direction: row;
