+++ src/Visualization/Components/Styles/NotificationPage.css
@@ -48,6 +48,9 @@
 }
 
 .main .content {
+	display: flex;
+	flex-direction: row;
+	justify-content: space-between;
 	padding: 10px;
 	height: 30px;
 	width: 1400px;
@@ -62,14 +65,8 @@
 	color: black;
 }
 
-.main .content h1 {
+.main .content .notification-time-span {
 	font-family: 'Lucida Console', 'Courier New', monospace;
 	color: black;
 	font-size: small;
-	max-width: 46px;
-	position: relative;
-	left: 1300px;
-	top: -7px;
-	border-bottom: 1px solid black;
-	border-right: 1px solid black;
 }
