+++ src/Visualization/Components/Styles/ProductPage.css
@@ -99,381 +99,3 @@
 	border-radius: 100%;
 	margin-bottom: 5px;
 }
-
-.product-page .bottom-section {
-	border: 1px solid #b6b6b6;
-	background-color: #e9e9e9;
-	border-radius: 5px;
-	margin-top: 30px;
-	padding: 1%;
-}
-
-.product-page .questions label {
-	font-family: 'Merriweather', sans-serif;
-}
-
-.product-page .questions input[type='text'] {
-	height: 30px;
-	width: 1500px;
-	border-radius: 5px;
-	margin-left: 15px;
-	margin-right: 15px;
-}
-
-.product-page .questions input[type='submit'] {
-	font-family: 'Merriweather Sans', sans-serif;
-	border-radius: 3px;
-	background-color: #7dd359;
-	border: 1px solid transparent;
-	cursor: pointer;
-	height: 34px;
-	width: 50px;
-}
-
-.product-page .questions input[type='submit']:hover {
-	background-color: rgb(35, 216, 133);
-}
-
-.product-page .old-questions {
-	border: 1px solid #b6b6b6;
-	margin-top: 20px;
-	padding: 1%;
-}
-
-.product-page .old-questions .title {
-	margin-left: 0px;
-	margin-bottom: 10px;
-	font-weight: bold;
-	text-transform: uppercase;
-	font-family: 'Merriweather Sans', sans-serif;
-}
-
-.product-page .old-questions .first-question {
-	border: 2px solid black;
-	border-radius: 5px;
-	padding: 0.5%;
-	margin-bottom: 10px;
-}
-
-.product-page .old-questions .second-question {
-	border: 2px solid black;
-	border-radius: 5px;
-	padding: 0.5%;
-	margin-bottom: 10px;
-}
-
-.product-page .old-questions .first-question label {
-	font-weight: bold;
-	border-bottom: 1px solid black;
-}
-
-.product-page .old-questions .second-question label {
-	font-weight: bold;
-	border-bottom: 1px solid black;
-}
-
-.product-page .old-questions .second-question .response {
-	border: 2px solid rgb(8, 137, 241);
-	padding: 0.25%;
-}
-
-.product-page .recomendations-section .product-page-showcase label {
-	font-family: 'Lucida Console', 'Courier New', monospace;
-	color: black;
-	width: 70%;
-	text-transform: uppercase;
-	font-size: 25px;
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-start;
-	margin-top: 50px;
-	margin-left: 100px;
-}
-
-.product-page .recomendations-section .product-page-showcase .product-showcase {
-	height: 245px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1 {
-	padding: 0.5%;
-	border: 2px solid #7dd359;
-	margin-bottom: 20px;
-	border-radius: 5px;
-	background-color: #c5c5c5;
-	height: 200px;
-	width: 450px;
-	position: absolute;
-	top: 1275px;
-	left: 230px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1:hover {
-	background-color: #b1b1b1;
-	cursor: pointer;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1
-	img {
-	height: 200px;
-	width: 200px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1
-	label {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-size: medium;
-	position: absolute;
-	top: -42px;
-	left: 118px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1
-	p {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-weight: lighter;
-	position: absolute;
-	top: 20px;
-	left: 220px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1
-	button {
-	padding: 2%;
-	font-family: 'Merriweather Sans', sans-serif;
-	border-radius: 4px;
-	background-color: #7dd359;
-	color: black;
-	position: inherit;
-	top: 170px;
-	left: 350px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product1
-	button:hover {
-	background-color: rgb(35, 216, 133);
-	cursor: pointer;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2 {
-	padding: 0.5%;
-	border: 2px solid #7dd359;
-	margin-bottom: 20px;
-	border-radius: 5px;
-	background-color: #c5c5c5;
-	height: 200px;
-	width: 450px;
-	position: absolute;
-	top: 1275px;
-	left: 730px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2
-	img {
-	height: 200px;
-	width: 200px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2:hover {
-	background-color: #b1b1b1;
-	cursor: pointer;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2
-	img {
-	height: 200px;
-	width: 200px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2
-	label {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-size: medium;
-	position: absolute;
-	top: -42px;
-	left: 118px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2
-	p {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-weight: lighter;
-	position: absolute;
-	top: 20px;
-	left: 220px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2
-	button {
-	padding: 2%;
-	font-family: 'Merriweather Sans', sans-serif;
-	border-radius: 4px;
-	background-color: #7dd359;
-	color: black;
-	position: inherit;
-	top: 170px;
-	left: 350px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product2
-	button:hover {
-	background-color: rgb(35, 216, 133);
-	cursor: pointer;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3 {
-	padding: 0.5%;
-	border: 2px solid #7dd359;
-	margin-bottom: 20px;
-	border-radius: 5px;
-	background-color: #c5c5c5;
-	height: 200px;
-	width: 450px;
-	position: absolute;
-	top: 1275px;
-	left: 1230px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3
-	img {
-	height: 200px;
-	width: 200px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3:hover {
-	background-color: #b1b1b1;
-	cursor: pointer;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3
-	img {
-	height: 200px;
-	width: 200px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3
-	label {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-size: medium;
-	position: absolute;
-	top: -42px;
-	left: 118px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3
-	p {
-	font-family: 'Merriweather Sans', sans-serif;
-	font-weight: lighter;
-	position: absolute;
-	top: 20px;
-	left: 220px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3
-	button {
-	padding: 2%;
-	font-family: 'Merriweather Sans', sans-serif;
-	border-radius: 4px;
-	background-color: #7dd359;
-	color: black;
-	position: inherit;
-	top: 170px;
-	left: 350px;
-}
-
-.product-page
-	.recomendations-section
-	.product-page-showcase
-	.product-showcase
-	.product3
-	button:hover {
-	background-color: rgb(35, 216, 133);
-	cursor: pointer;
-}
